Add comprehensive test coverage and refactor extension into modular structure.
