+++ .github/workflows/test.yml
@@ -9,7 +9,7 @@ on:
 jobs:
   test:
     if: github.event_name == 'push' || github.event.pull_request.head.repo.full_name != github.repository
-    name: Linter
+    name: Test
     runs-on: ubuntu-latest
     timeout-minutes: 10
     steps:
@@ -30,3 +30,6 @@ jobs:
 
     - name: Run the build
       run: npm run build
+
+    - name: Run tests
+      run: xvfb-run -a npm test
